feat(FilesTableSort): refactor to its own component
